requirements.txt
torch
tensorflow
keras
numpy
scipy
scikit-learn
dask
pandas
matplotlib
seaborn
umap-learn
lightning
torchvision
pytables
